Add files via upload
Restore this corrupted PNG image by changing one byte in the file. No encryption, histogram or lsb techniques used. The aim of this exercise is to determine whether future LLMs will be able to parse image data to detect anomalies or changes in byte order and structure to uncover concealed data.
This method is highly resistant to current Steganalysis tools as a person would have to 'guess' which of the  152 371 bytes is the incorrect one.